+++ noodles-bam/src/reader/record/data.rs
@@ -1,3 +1,5 @@
+//! BAM record data field reader.
+
 pub mod field;
 
-pub use self::field::read_field;
+pub(crate) use self::field::read_field;